refactor(api): migrate /api/news to Zod via defineRoute
